Fix and Improvements
I have cleaned up some of the Header code in the css file from index.css.

and I have added some line's in by the background image inside of the index.html file so that the image can be fixed and changed the height of the image to 440px so that its able to fit nicely.